String params, gui start(?)
Overloaded Rubik.Rotate() such that it may now accept strings as parameters. I'm also intending to make a function that can parse a long string of moves in Rubik notation, so that we can simply pass those moves into Rotate() one at a time.
Also, I started doing a shenanigan with the GUI, but it's a little... off. If you look at the "Game" screen, and then hit Play while "Maximize on play" is active, you'll see what I mean.