Smooth
For the record, it was working in class--I was just pushign the wrong button.
The x, y, and z inputs now move smoothly (without spinning endlessly), and the user can't input anything while any part of the cube is turning. If it needs to be sped up, there's a public float variable called "Turn Speed" (... I think that's what I called it).
Now, I just need to do the same with the rotation of the sides of the cube.